+++ src/list.h
@@ -1,6 +1,8 @@
 #ifndef _LIST_H_
 #define _LIST_H_
 
+#include <stddef.h>
+
 typedef struct list_t 
 {
 	struct list_t* next;
@@ -11,7 +13,7 @@ list_t* list_create(void* data);
 list_t* list_node_at(list_t* head, int id);
 void* list_get(list_t* head, int id);
 void list_free(list_t** head);
-int list_length(list_t* head);
+size_t list_length(list_t* head);
 void list_add_front(list_t** head, void* data);
 void list_add_back(list_t** head, void* data);
 void list_remove(list_t** head, void* data);
